requirements.txt
streamlit
langchain
langchain_community
langchain_groq
langchain_huggingface
faiss-cpu
python-dotenv
pypdf
pymupdf
faiss-cpu